Creating the hero for my website.
First in html I layed out a <h1> for the header
of the hero part of the page, then I added <p>
under the header and after that i added a sign up
button. lastly I added a "placeholer for a image"
witch it actually just a grey box.

After this I header over to css and made the hero
elements a column flexbox and set up the header
and text as well as the button in a column
layout. For the "placeholer" I just made a grey
box and since it's not in the hero section and
not the content it went to the right of the text.